authen & author for user manage admin

+++ src/java/controller/AccountController.java
@@ -7,7 +7,6 @@ package controller;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
-import static controller.ReportController.reportService;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -16,6 +15,7 @@ import javax.mail.MessagingException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import middleware.VerifyRequest;
 import model.User;
 import service.AccountService;
 import serviceImplement.AccountServiceImplement;
@@ -44,7 +44,7 @@ public class AccountController {
             System.out.println("new user: " + userId);
             if (userId != -1) {
                 status.put("status", "success");
-                status.put("roll", accountService.getUserRoll(email));
+                status.put("role", accountService.getUserRole(email));
 
                 // store email in http session
                 HttpSession session = request.getSession();
@@ -130,23 +130,29 @@ public class AccountController {
     }
 
     public static void getUsers(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        ArrayList<User> users = accountService.getUsers();
-        String json = new Gson().toJson(users);
-        System.out.println(json);
-        response.setContentType("application/json");
-        response.setCharacterEncoding("UTF-8");
-        response.getWriter().write(json);
+        if (VerifyRequest.verifyUserManageRequest(request, response)) {
+            ArrayList<User> users = accountService.getUsers();
+            String json = new Gson().toJson(users);
+            System.out.println(json);
+            response.setContentType("application/json");
+            response.setCharacterEncoding("UTF-8");
+            response.getWriter().write(json);
+        }
     }
 
     public static void banUser(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int uid = Integer.parseInt(request.getParameter("uid"));
-        accountService.banUser(uid);
-        response.sendRedirect("http://localhost:8080/quora-admin-client/admin/user-manage/user/user.jsp");
+        if (VerifyRequest.verifyUserManageRequest(request, response)) {
+            int uid = Integer.parseInt(request.getParameter("uid"));
+            accountService.banUser(uid);
+            response.sendRedirect("http://localhost:8080/quora-admin-client/admin/user-manage/user/user.jsp");
+        }
     }
 
     public static void deleteUser(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int uid = Integer.parseInt(request.getParameter("uid"));
-        accountService.deleteUser(uid);
-        response.sendRedirect("http://localhost:8080/quora-admin-client/admin/user-manage/user/user.jsp");
+        if (VerifyRequest.verifyUserManageRequest(request, response)) {
+            int uid = Integer.parseInt(request.getParameter("uid"));
+            accountService.deleteUser(uid);
+            response.sendRedirect("http://localhost:8080/quora-admin-client/admin/user-manage/user/user.jsp");
+        }
     }
 }
