+++ src/java/controller/ReportController.java
@@ -10,6 +10,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import middleware.VerifyRequest;
 import model.Report;
 import service.ReportService;
 import serviceImplement.ReportServiceImplement;
@@ -23,18 +24,21 @@ public class ReportController {
     static ReportService reportService = new ReportServiceImplement();
 
     public static void getReports(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        ArrayList<Report> reports = reportService.getReports();
-        String json = new Gson().toJson(reports);
-        System.out.println(json);
-        response.setContentType("application/json");
-        response.setCharacterEncoding("UTF-8");
-        response.getWriter().write(json);
+        if (VerifyRequest.verifyUserManageRequest(request, response)) {
+            ArrayList<Report> reports = reportService.getReports();
+            String json = new Gson().toJson(reports);
+            System.out.println(json);
+            response.setContentType("application/json");
+            response.setCharacterEncoding("UTF-8");
+            response.getWriter().write(json);
+        }
     }
 
     public static void deleteReport(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int aid = Integer.parseInt(request.getParameter("aid"));
-        reportService.deleteReport(aid);
-        response.sendRedirect("http://localhost:8080/quora-admin-client/admin/user-manage/report/report.jsp");
+        if (VerifyRequest.verifyUserManageRequest(request, response)) {
+            int aid = Integer.parseInt(request.getParameter("aid"));
+            reportService.deleteReport(aid);
+            response.sendRedirect("http://localhost:8080/quora-admin-client/admin/user-manage/report/report.jsp");
+        }
     }
-
 }
